+++ GeneratorLibrary/Models/HydrographicCoverage.cs
@@ -2,7 +2,7 @@
 {
     public record HydrographicCoverage
     {
-        public float Coverage { get; set; }
+        public double Coverage { get; set; }
         public List<string> Composition { get; set; } = new List<string>();
     }
 }
